# src/SvarOMatic.Backend/wwwroot/swagger-original.json
{
  "openapi" : "3.0.0",
  "info" : {
    "title" : "Swagger - Svar-o-Matic",
    "description" : "This is the backend for a solution that allows employees of Miles AS to efficiently match incoming project requests with existing consultant expertise.\n\nSome useful links:\n- [The Svar-o-Matic repository](https://github.com/miles-no/svar-o-matic)",
    "termsOfService" : "http://swagger.io/terms/",
    "contact" : {
      "email" : "drift@miles.no"
    },
    "license" : {
      "name" : "Apache 2.0",
      "url" : "http://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version" : "1.0.11"
  },
...
  "servers" : [ {
    "url" : "http://localhost:5090"
  } ],
...
  "paths" : {
...
    "/svar/{consultantId}/{inquiryId}/{kravId}/accept" : {
      "post" : {
        "tags" : [ "svar" ],
        "summary" : "Accept a svar to a krav",
        "description" : "Accept a svar to a krav",
        "operationId" : "acceptSvar",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the deletion is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "kravId",
          "in" : "path",
          "description" : "Id of the krav that the operation is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
    "/svar/{consultantId}/{inquiryId}/{kravId}/edit" : {
      "post" : {
        "tags" : [ "svar" ],
        "summary" : "Manually edit a svar to a krav",
        "description" : "Manually edit a svar to a krav",
        "operationId" : "editSvar",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the deletion is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "kravId",
          "in" : "path",
          "description" : "Id of the krav that the operation is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "requestBody" : {
          "description" : "Data about the svar",
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/Svar"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/Svar"
              }
            },
            "application/x-www-form-urlencoded" : {
              "schema" : {
                "$ref" : "#/components/schemas/Svar"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
    "/svar/{consultantId}/{inquiryId}/{kravId}/generateSvar" : {
      "post" : {
        "tags" : [ "svar" ],
        "summary" : "Generate answer for a Krav",
        "description" : "Generate answer for a Krav",
        "operationId" : "generateAnswerForKrav",
        "parameters" : [ {
          "name" : "consultantId",
          "in" : "path",
          "description" : "Id of the consultant that the addition is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "inquiryId",
          "in" : "path",
          "description" : "Id of the forespørsel that the deletion is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        }, {
          "name" : "kravId",
          "in" : "path",
          "description" : "Id of the krav that the request is related to",
          "required" : true,
          "style" : "simple",
          "explode" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int64"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "Successful operation",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              },
              "application/xml" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "405" : {
            "description" : "Invalid input"
          }
        },
        "security" : [ {
          "svaromatic_inquiry_auth" : [ "write:inquiry", "read:inquiry" ]
        } ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "Svar" : {
        "type" : "object",
        "properties" : {
          "kravId" : {
            "type" : "integer",
            "format" : "int64"
          },
          "text" : {
            "type" : "string"
          },
          "generatedOn" : {
            "type" : "string",
            "format" : "date-time"
          }
        }
      },
      "ApiResponse" : {
        "type" : "object",
        "properties" : {
          "code" : {
            "type" : "integer",
            "format" : "int32"
          },
          "type" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          }
        },
        "example" : {
          "code" : 0,
          "type" : "type",
          "message" : "message"
        },
        "xml" : {
          "name" : "##default"
        }
      }
    },
...
    "securitySchemes" : {
      "svaromatic_docs_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify docs for a consultant",
              "read:cvs" : "read your docs"
            }
          }
        }
      },
      "svaromatic_inquiry_auth" : {
        "type" : "oauth2",
        "flows" : {
          "implicit" : {
            "authorizationUrl" : "https://coming-soon.swagger.io/oauth/authorize",
            "scopes" : {
              "write:cvs" : "modify inquiries for a consultant",
              "read:cvs" : "read inquiries"
            }
          }
        }
      },
      "api_key" : {
        "type" : "apiKey",
        "name" : "api_key",
        "in" : "header"
      }
    }
  }
}